+++ lib/auth/storage/storage_unix.go
@@ -1,5 +1,5 @@
-//go:build windows
-// +build windows
+//go:build !windows
+// +build !windows
 
 /*
  * Teleport
@@ -19,25 +19,47 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package state
+package storage
 
 import (
 	"context"
 
 	"github.com/gravitational/trace"
 
-	"github.com/gravitational/teleport/lib/backend/memory"
+	"github.com/gravitational/teleport/lib/backend/kubernetes"
+	"github.com/gravitational/teleport/lib/backend/lite"
 )
 
 // NewProcessStorage returns a new instance of the process storage.
 func NewProcessStorage(ctx context.Context, path string) (*ProcessStorage, error) {
-	m, err := memory.New(memory.Config{
-		Context:   ctx,
+	if path == "" {
+		return nil, trace.BadParameter("missing parameter path")
+	}
+
+	litebk, err := lite.NewWithConfig(ctx, lite.Config{
+		Path:      path,
 		EventsOff: true,
+		Sync:      lite.SyncFull,
 	})
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
 
-	return &ProcessStorage{BackendStorage: m, stateStorage: m}, nil
+	// identityStorage holds the storage backend for identity and state.
+	// if the agent is running in Kubernetes it's replaced by kubernetes secret storage
+	var identityStorage stateBackend = litebk
+
+	// if running in a K8S cluster and required env vars are available
+	// the agent will automatically switch state storage from local
+	// sqlite into a Kubernetes Secret.
+	if kubernetes.InKubeCluster() {
+		kubeStorage, err := kubernetes.New()
+		if err != nil {
+			return nil, trace.Wrap(err)
+		}
+
+		identityStorage = kubeStorage
+	}
+
+	return &ProcessStorage{BackendStorage: litebk, stateStorage: identityStorage}, nil
 }
